requirements/base.txt
boto3==1.26.118
dj-rest_auth>=3.0.0,<3.1.0
django-anymail==9.1
django-cors-headers==3.14.0
django-filter==23.2
django-simple-history==3.3.0
djangorestframework==3.14.0,<3.15.0
djangorestframework-simplejwt>=5.2.2,<5.3.0
drf-spectacular>=0.26.0,<0.27.0
Faker==18.7.0
pre-commit==3.3.3
psycopg2-binary==2.9.6
PyJWT==2.7.0
python-dotenv==1.0.0
requests==2.30.0
